requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==9.5.0
ultralytics==8.2.20
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0